[Bugfix] Jar file curl failure in exploit-server
The download for the JNDI-Exploit-Kit was done using a curl command in
the dockerfile. A recent commit in the upstream GH repo moved this file
from https://github.com/pimps/JNDI-Exploit-Kit/raw/2ca240fe3d552c7fb2c3b04e9a457e8c8ae98447/target/JNDI-Exploit-Kit-1.0-SNAPSHOT-all.jar to https://github.com/pimps/JNDI-Exploit-Kit/raw/master/JNDI-Exploit-Kit.jar resulting in a 404 error page being pulled in place of the JAR file.

This resulted in errors in the logs for the exploit-server and the JNDI
Injection attack not working.

Fixed by committinng the files to our repo and pinning. Tested with the
new (updated) JNDI-Exploit-Kit.jar file and this is working.